// types/ShowTop.ts
export interface Album {
    album_type: string;
    total_tracks: number;
    available_markets: string[];
    external_urls: {
        spotify: string;
    };
    href: string;
    id: string;
    images: {
        url: string;
        height: number;
        width: number;
    }[];
    name: string;
    release_date: string;
    release_date_precision: string;
    type: string;
    uri: string;
    artists: ArtistForTrack[];
}
  
export interface ArtistForTrack {
    external_urls: {
        spotify: string;
    };
    href: string;
    id: string;
    name: string;
    type: string;
    uri: string;
}
  
export interface Track {
    album: Album;
    artists: ArtistForTrack[];
    available_markets: string[];
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    external_ids: {
        isrc: string;
    };
    external_urls: {
        spotify: string;
    };
    href: string;
    id: string;
    name: string;
    popularity: number;
    preview_url: string;
    track_number: number;
    type: string;
    uri: string;
    is_local: boolean;
}

export interface Artist {
    external_urls: {
        spotify: string
    }
    followers: {
        href: any
        total: number
    }
    genres: string[]
    href: string
    id: string
    images: {
        url: string
        height: number
        width: number
    }[]
    name: string
    popularity: number
    type: string
    uri: string
}